pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-trading-bot"
version = "2.0.0"
description = "BTC Trading Bot with FastAPI, Google Sheets, Coinbase WebSocket, and Telegram"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "BTC Trading Bot Team"},
]
keywords = ["bitcoin", "trading", "fastapi", "coinbase", "telegram"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv>=1.0",
    "gspread",
    "google-auth",
    "google-auth-oauthlib",
    "websocket-client",
    "coinbase-advanced-py",
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "jinja2",
    "aiofiles",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio", 
    "httpx",
    "black",
    "flake8",
    "isort",
    "mypy",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/your-org/btc-trading-bot"
Issues = "https://github.com/your-org/btc-trading-bot/issues"

[project.scripts]
btc-trading-bot = "app.main:app"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "gspread.*",
    "coinbase.*",
    "websocket.*",
]
ignore_missing_imports = true